pyproject.toml
[project]
name = "ai-methods-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "langchain>=0.3.25",
  "langchain-gigachat>=0.3.10",
  "langgraph>=0.4.3",
  "python-dotenv>=1.1.0",
  "python-telegram-bot>=22.1",
]
